requirements.txt
Django>=5.2.7
daphne>=4.1.0
channels>=4.1.0
channels-redis>=4.2.0
djangorestframework>=3.15.0
django-cors-headers>=4.4.0
python-decouple>=3.8
redis>=5.2.0
psycopg2-binary>=2.9.9
whitenoise>=6.8.0
gunicorn>=23.0.0
celery>=5.4.0
Pillow>=10.0.0
